+++ tests/test_methods.py
@@ -7,7 +7,6 @@ Created on 19 Dec 2020
 # pylint: disable=missing-class-docstring, missing-function-docstring
 
 import unittest
-from subprocess import PIPE, run
 
 from semuadmin_sandpit.calculate import Calculate
 
@@ -40,15 +39,6 @@ class StaticTest(unittest.TestCase):
                 c.calc("divide", 2.5, 3.5),
                 0.7142857142857143,
             )
-
-    def testcli(self):  # must be installed via pip first
-        res = run(
-            ["calculate", "--function", "add", "--arg1", "2.5", "--arg2", "3.5"],
-            stdout=PIPE,
-            check=False,
-        )
-        res = res.stdout.decode("utf-8").strip("\r\n")
-        self.assertEqual(res, "6.0")
 
 
 if __name__ == "__main__":
